#6 - Refactor: 글 관련 코드 스타일 변경
