+++ 2Week_Mission/src/main/java/com/ll/exam/ebooks/app/base/initData/InitDataBefore.java
@@ -3,7 +3,7 @@ package com.ll.exam.ebooks.app.base.initData;
 import com.ll.exam.ebooks.app.member.form.JoinForm;
 import com.ll.exam.ebooks.app.member.entity.Member;
 import com.ll.exam.ebooks.app.member.service.MemberService;
-import com.ll.exam.ebooks.app.post.dto.PostForm;
+import com.ll.exam.ebooks.app.post.form.PostForm;
 import com.ll.exam.ebooks.app.post.entity.Post;
 import com.ll.exam.ebooks.app.post.service.PostService;
